feat: add tanstacck query, create postspage, clean code (#101)
