+++ client/package.json
@@ -5,6 +5,8 @@
   "private": true,
   "dependencies": {
     "@emailjs/browser": "^3.12.1",
+    "@tanstack/react-query": "^5.51.23",
+    "@tanstack/react-query-devtools": "^5.51.23",
     "axios": "^1.3.4",
     "quill": "^1.3.7",
     "react": "^18.2.0",
