+++ client/src/api/api.js
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const api = axios.create({
+  baseURL: process.env.REACT_APP_SERVER_URL || 'http://localhost:5005',
+});
+
+export { api };
