+++ client/src/services/index.js
@@ -1,0 +1,1 @@
+export * as postActions from './actions';
